[SECURITY-925] Fold tests back into normal files
